auth.service + login component
