--- client/src/components/Services/users.service.js
@@ -1,8 +1,0 @@
-import firebase from 'firebase'
-
-export default {
-  getCurrentUser () {
-  // check if user is logged
-    return firebase.auth().currentUser
-  }
-}
